Updated entity relationships to use lazy loading for improved performance; added staffId to BillingDTO and updated related entities.
I still have a lot to do, namely: making methods in each service class for each entity for converting DTOs to entities and vice versa, testing the JWT and other security stuff I implemented if it works (it probably won't), and resting for the upcoming slog that is "working on 3 capstone projects at the same time". Help me, I might go insane.