Moved @NotBlank and @Pattern to UserDTO; Increased length of password in UserEntity
Oversight on my part, trying to use /admin/register and putting in 8-18 characters would give an error since before the changes, UserEntity would only allow a limit of 18 characters for the password. Thing is, this is after it has been hashed using BCrypt, so it's increased to 60 characters. Hopefully this should fix the issue.